requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
transformers==4.35.2
torch>=2.2.0
requests==2.31.0
reportlab==4.0.7
python-dotenv==1.0.0
Pillow==10.1.0
PyJWT==2.8.0
